Ispravlyau oshibki

+++ scheduler/src/main/java/ru/msu/cmc/scheduler/DAO/AuditoriumDAO.java
@@ -6,6 +6,6 @@ import ru.msu.cmc.scheduler.models.Auditorium;
 
 import java.util.List;
 
-public interface AuditoriumDAO extends CommonDAO<Auditorium, Long> {
+public interface AuditoriumDAO extends CommonDAO<Auditorium, Integer> {
 
 }
