+++ scheduler/src/main/java/ru/msu/cmc/scheduler/DAO/StudentDAO.java
@@ -1,12 +1,10 @@
 package ru.msu.cmc.scheduler.DAO;
 
-import lombok.Builder;
-import lombok.Getter;
 import ru.msu.cmc.scheduler.models.Student;
 
 import java.util.List;
 
-public interface StudentDAO extends CommonDAO<Student, Long> {
+public interface StudentDAO extends CommonDAO<Student, Integer> {
 
     List<Student> getAllStudentsByName(String studentName);
